new data model extended to CC

+++ ced/src/main/java/cnuphys/ced/cedview/rtpc/RTPCView.java
@@ -290,9 +290,9 @@ public class RTPCView extends CedXYView implements ChangeListener {
 					short component =  (short)(cm1 + 1);
 					byte layer = (byte)(lm1 + 1);
 					setRect(container, component, layer, hr, p0, p1);
-					
+
 					double hitCount = counts[cm1][lm1];
-					double fract = (maxHit == 0) ? 0 : (((double) hitCount) / maxHit);
+					double fract = (maxHit == 0) ? 0 : ((hitCount) / maxHit);
 					Color color = AccumulationManager.getInstance().getColor(getColorScaleModel(), fract);
 					g.setColor(color);
 					g.fillRect(hr.x, hr.y, hr.width, hr.height);
